adding auto msbuild sdk setup action
