+++ Dockerfile.msbuild
@@ -1,0 +1,9 @@
+FROM mcr.microsoft.com/dotnet/framework/sdk:4.7.2
+
+WORKDIR /app
+
+COPY FontDumpTruck/ .
+RUN nuget restore .
+RUN msbuild /p:Configuration=Release /p:OutputPath=C:\app\out\
+
+ENTRYPOINT ["/app/out/FontDumpTruck.exe"]
