+++ action.yml
@@ -1,0 +1,10 @@
+# action.yml
+name: '.NetFramework Action Template'
+description: 'Template for .Net Framework Action'
+author: 'Gumpanat Keardkeawfa <https://github.com/auycro>'
+branding:
+  icon: 'book-open'
+  color: 'blue'
+runs:
+  using: 'docker'
+  image: 'Dockerfile'
